Refactored worker execution logic to be true async suspension
Add WorkflowRegistration and WorkflowScheduling implementations to handle workflow activation, cached re-activation, checkpointing, history/events, timers, child scheduling and event waits. Persist and use SchedulingMetadata in task rows: updated DB queries and row decoding to read/write scheduling_metadata instead of using opaque header JSON. Improve JSON helpers by reusing a shared ByteBufferAllocator/JSONEncoder/JSONDecoder, adding a string decode helper, and tightening error handling. Fix TaskResultSnapshot.decode to use the new JSON helpers. Update SQL/schema usage to include scheduling_metadata and adjust enqueue/insert queries accordingly.
