+++ Sources/Strand/Worker.swift
@@ -127,1560 +127,6 @@ public struct WorkerOptions: Sendable {
     }
 }
 
-// MARK: - Workflow._makeToken() default implementation
-//
-// Lives here (not Workflow.swift) because it closes over _WorkflowTaskCache<Self>
-// and WorkflowRegistration<Self>, both defined in this file.
-// The cache is created once per workflow TYPE at registration time and shared
-// across every activation of that type, enabling the cached-Task model.
-extension Workflow {
-    public static func _makeToken() -> _WorkflowToken {
-        let cache = _WorkflowTaskCache<Self>()
-        return _WorkflowToken(name: workflowName, preferredQueue: nil) { claimed, exec in
-            try await WorkflowRegistration<Self>().activate(
-                claimed: claimed,
-                exec: exec,
-                cache: cache
-            )
-        }
-    }
-}
-
-// MARK: - _WorkflowTaskCache
-
-/// Per-workflow-type cache of live handler Tasks.
-///
-/// One instance per workflow type, created at token-registration time and captured
-/// by the token closure. The handler `Task` stays parked on its continuations
-/// between activations, consuming no thread. On re-activation the worker delivers
-/// real results via the Resume API and calls `drain()` to continue from where
-/// the handler paused.
-///
-/// `@unchecked Sendable`: access to the dictionary is serialised by the `Mutex`;
-/// access to the cached executor is serialised by Postgres
-/// (`FOR UPDATE SKIP LOCKED` prevents two workers from claiming the same run).
-final class _WorkflowTaskCache<W: Workflow>: @unchecked Sendable {
-
-    struct CachedState: @unchecked Sendable {
-        let executor: StrandWorkflowExecutor
-        let task: Task<Void, Never>
-        let handlerResult: ArcBox<Result<W.Output, Error>?>
-        let stateBox: ArcBox<W>
-        let activation: _WorkflowActivation<W>
-    }
-
-    private let _mutex = Mutex<[UUID: CachedState]>([:])
-
-    func get(_ taskID: UUID) -> CachedState? {
-        _mutex.withLock { $0[taskID] }
-    }
-
-    func set(_ taskID: UUID, _ state: CachedState) {
-        _mutex.withLock { $0[taskID] = state }
-    }
-
-    func remove(_ taskID: UUID) {
-        _ = _mutex.withLock { $0.removeValue(forKey: taskID) }
-    }
-
-    /// Cancel every cached Task — called from a cancellation handler so that
-    /// worker shutdown also cleans up Tasks that are parked mid-activation.
-    func cancelAll() {
-        _mutex.withLock { states in
-            for (_, state) in states {
-                state.executor.cancelPending()
-                state.executor.drain()
-                state.task.cancel()
-            }
-            states.removeAll()
-        }
-    }
-}
-
-// MARK: - WorkflowRegistration
-
-/// Internal activation implementation for `Workflow`-conforming types.
-struct WorkflowRegistration<W: Workflow>: Sendable {
-
-    /// Runs one activation of workflow type `W` for the given claimed task.
-    ///
-    /// Returns the encoded result `ByteBuffer` on successful completion, or `nil`
-    /// when the activation suspended (handler Task left parked on a continuation).
-    func activate(
-        claimed: ClaimedTask,
-        exec: _WorkerExec,
-        cache: _WorkflowTaskCache<W>
-    ) async throws -> ByteBuffer? {
-        // Wrap the whole activation so that worker shutdown (Task cancellation)
-        // also tears down any cached handler Task for this workflow instance.
-        try await withTaskCancellationHandler {
-            try await _activate(claimed: claimed, exec: exec, cache: cache)
-        } onCancel: {
-            cache.cancelAll()
-        }
-    }
-
-    private func _activate(
-        claimed: ClaimedTask,
-        exec: _WorkerExec,
-        cache: _WorkflowTaskCache<W>
-    ) async throws -> ByteBuffer? {
-
-        // ── Cached path: Task is alive and parked on its continuations ────────────
-        if let cached = cache.get(claimed.taskID) {
-            return try await resumeActivation(claimed: claimed, exec: exec, cached: cached, cache: cache)
-        }
-
-        // ── Fresh path: first activation (or after a crash / cache miss) ─────────
-
-        // ── 1. Checkpoint cache ───────────────────────────────────────────────────
-        let checkpointRows = try await Queries.getCheckpointStates(
-            on: exec.postgres,
-            taskID: claimed.taskID,
-            runID: claimed.runID,
-            logger: exec.logger
-        )
-        var checkpointCache: [Int: ByteBuffer] = [:]
-        for row in checkpointRows { checkpointCache[row.seqNum] = row.stateBuffer }
-
-        // ── 2. Workflow state ────────────────────────────────────────────────────
-        let storedStateBuf = try await WorkflowStateQueries.loadState(
-            on: exec.postgres,
-            taskID: claimed.taskID,
-            namespaceID: exec.namespace,
-            logger: exec.logger
-        )
-        var workflowState: W
-        if let buf = storedStateBuf {
-            // Restore the struct exactly as it was at the end of the previous activation.
-            workflowState = try JSON.decode(W.self, from: buf)
-        } else {
-            // First activation — no stored state yet. Use the protocol-required
-            // `init()` so non-optional stored properties work without any
-            // Codable boilerplate (no custom `init(from:)` needed).
-            workflowState = W()
-        }
-
-        // ── History sequence ──────────────────────────────────────────────────
-        // Use historySeq == 1 (no events written yet) as the "first activation" signal
-        // instead of `storedStateBuf == nil`. `workflow_state` is only saved on signal
-        // delivery or completion, so storedStateBuf is nil on every pure activity-wait
-        // re-activation — causing a duplicate WORKFLOW_STARTED event.
-        var historySeq = try await WorkflowStateQueries.nextHistorySeq(
-            on: exec.postgres,
-            taskID: claimed.taskID,
-            logger: exec.logger
-        )
-        let isFirstActivation = historySeq == 1
-        if isFirstActivation {
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .workflowStarted,
-                eventData: nil,
-                logger: exec.logger
-            )
-            historySeq += 1
-        }
-
-        // ── 3. Signals ────────────────────────────────────────────────────────────
-        let signals = try await WorkflowStateQueries.loadPendingSignals(
-            on: exec.postgres,
-            taskID: claimed.taskID,
-            namespaceID: exec.namespace,
-            logger: exec.logger
-        )
-        for signal in signals {
-            try workflowState.handleSignal(name: signal.name, payload: signal.payload)
-        }
-        if !signals.isEmpty {
-            for signal in signals {
-                let sigData = try? JSON.encode(["name": signal.name])
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .signalReceived,
-                    eventData: sigData,
-                    logger: exec.logger
-                )
-                historySeq += 1
-            }
-            let postSignalBuf = try JSON.encode(workflowState)
-            try await WorkflowStateQueries.saveState(
-                on: exec.postgres,
-                taskID: claimed.taskID,
-                namespaceID: exec.namespace,
-                stateBuffer: postSignalBuf,
-                logger: exec.logger
-            )
-            try await WorkflowStateQueries.deleteSignals(
-                on: exec.postgres,
-                ids: signals.map { $0.id },
-                logger: exec.logger
-            )
-        }
-
-        // ── 4. Pre-load results the executor needs for fast-path replay ───────────
-        let executor = StrandWorkflowExecutor()
-
-        // 4a. Terminal child activities — runActivity / runChildWorkflow returns or
-        //     throws immediately for these (COMPLETED = return, FAILED = throw).
-        let completedChildren = try await WorkflowStateQueries.loadCompletedChildActivities(
-            on: exec.postgres,
-            parentTaskID: claimed.taskID,
-            logger: exec.logger
-        )
-        executor.resolveCompleted(completedChildren)
-        for (seqNum, _, _, state, kind, name) in completedChildren
-        where kind == .workflow && state == .completed {
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .childWorkflowCompleted,
-                // seq_num is a String for consistency with CHILD_WORKFLOW_STARTED
-                eventData: try? JSON.encode(["workflow": name, "seq_num": String(seqNum)]),
-                logger: exec.logger
-            )
-            historySeq += 1
-        }
-
-        // ── 5. Build activation context ────────────────────────────────────────────
-        // Capture wall-clock time once here so WorkflowContext.activationTime
-        // is stable for the entire duration of this activation.
-        let activationTime = Date()
-        let claimTimeoutSecs = Int(exec.options.claimTimeout.components.seconds)
-        let stateBox = ArcBox(workflowState)
-        let activation = _WorkflowActivation<W>(
-            taskUUID: claimed.taskID,
-            runUUID: claimed.runID,
-            taskName: claimed.taskName,
-            queueName: exec.queue,
-            attempt: claimed.attempt,
-            claimTimeoutSeconds: claimTimeoutSecs,
-            wakeEvent: claimed.wakeEvent,
-            eventPayload: claimed.eventPayloadBuffer,
-            headers: claimed.headers,
-            postgres: exec.postgres,
-            logger: exec.logger,
-            executor: executor,
-            stateBox: stateBox,
-            checkpointCache: checkpointCache,
-            namespace: exec.namespace,
-            activationTime: activationTime
-        )
-        let context = WorkflowContext<W>(activation: activation)
-        let input = try JSON.decode(W.Input.self, from: claimed.paramsBuffer)
-
-        // ── 6. Run handler on the executor ─────────────────────────────────────────
-        // All WorkflowContext operations (runActivity, sleep, waitForEvent, uuid, random)
-        // emit WorkflowCommands into executor.pendingCommands — zero DB I/O inside the
-        // handler. When a task suspends (pending activity, timer, event), it parks a
-        // CheckedContinuation in the executor.
-        //
-        // ArcBox<Result?> shares the handler outcome between the Task closure and the
-        // outer activation code through the serial drain() synchronisation point.
-        // No concurrent mutations occur in practice: the closure runs only during
-        // drain(), which returns before the outer code reads handlerResult.
-        let handlerResult = ArcBox<Result<W.Output, Error>?>(nil)
-        let handlerTask = Task(executorPreference: executor) {
-            do {
-                let output = try await stateBox.value.run(context: context, input: input)
-                handlerResult.value = .success(output)
-            } catch {
-                handlerResult.value = .failure(error)
-            }
-        }
-
-        // Run the handler on the serial executor until it either completes or
-        // every in-flight task has parked its continuation (pending activity,
-        // timer, or event wait).
-        executor.drain()
-
-        // ── Condition fast-path ─────────────────────────────────────
-        // A signal or state change may have arrived before the handler reached
-        // context.condition — evaluate immediately rather than parking the run.
-        // Each resume may unblock more of the handler and emit new commands,
-        // so drain after each satisfied condition until none remain.
-        while executor.evaluateAndResumeFirstSatisfiedCondition() {
-            executor.drain()
-        }
-
-        // ── Local activity execution loop ─────────────────
-        // Local activities run in-process with no DB task row or queue dispatch.
-        // After each batch the executor flushes buffered commands and re-evaluates
-        // any pending conditions; break early if the handler has already finished.
-        localActivityLoop: while !executor.localActivityEntries.isEmpty {
-            let entries = executor.localActivityEntries  // snapshot
-            for (id, entry) in entries {
-                guard let runner = exec.localActivityLookup[entry.name] else {
-                    executor.failLocalActivity(
-                        id: id,
-                        error: StrandError.unknownTask(name: entry.name)
-                    )
-                    continue
-                }
-                do {
-                    let result = try await runner(entry.input, exec, claimed.taskID)
-                    try await Queries.setCheckpointState(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seqNum: entry.seqNum,
-                        name: entry.name,
-                        stateBuffer: result,
-                        runID: claimed.runID,
-                        extendClaimBySeconds: claimTimeoutSecs,
-                        logger: exec.logger
-                    )
-                    activation.cacheCheckpoint(seqNum: entry.seqNum, buffer: result)
-                    executor.resolveLocalActivity(id: id, result: result)
-                } catch {
-                    executor.failLocalActivity(id: id, error: error)
-                }
-            }
-            executor.drain()
-            while executor.evaluateAndResumeFirstSatisfiedCondition() { executor.drain() }
-            if handlerResult.value != nil { break localActivityLoop }
-        }
-
-        // ── 7. Apply commands to Postgres ─────────────────────────────────────────
-        // Split commands by type and apply them atomically.
-        let commands = executor.pendingCommands
-
-        // Non-suspending: writeCheckpoint, timerFired, eventReceived.
-        // Write these before anything else so they're durable regardless of what follows.
-        let checkpointWrites = commands.compactMap {
-            cmd -> (seqNum: Int, name: String?, state: ByteBuffer)? in
-            if case .writeCheckpoint(let seqNum, let name, let value) = cmd {
-                return (seqNum: seqNum, name: name, state: value)
-            }
-            return nil
-        }
-        if !checkpointWrites.isEmpty {
-            try await Queries.batchSetCheckpoints(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                runID: claimed.runID,
-                checkpoints: checkpointWrites,
-                extendClaimBySeconds: claimTimeoutSecs,
-                logger: exec.logger
-            )
-            for (seqNum, _, value) in checkpointWrites {
-                activation.cacheCheckpoint(seqNum: seqNum, buffer: value)
-            }
-        }
-
-        // Replay fast-path history events (TIMER_FIRED, EVENT_RECEIVED).
-        for cmd in commands {
-            switch cmd {
-            case .timerFired(_):
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .timerFired,
-                    eventData: nil,
-                    logger: exec.logger
-                )
-                historySeq += 1
-            case .eventReceived(let eventName):
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .eventReceived,
-                    eventData: try? JSON.encode(["event_name": eventName]),
-                    logger: exec.logger
-                )
-                historySeq += 1
-            default:
-                break
-            }
-        }
-
-        // Decide on the terminal state of this activation.
-        switch handlerResult.value {
-
-        case .success(let output):
-            // Handler completed — persist state and return the encoded result.
-            // The caller (runTask) will write COMPLETED to the run.
-            cache.remove(claimed.taskID)
-            handlerTask.cancel()
-            executor.cancelPending()
-            executor.drain()  // flush any residual jobs from cancelPending()
-            let finalStateBuf = try JSON.encode(stateBox.value)
-            try await WorkflowStateQueries.saveState(
-                on: exec.postgres,
-                taskID: claimed.taskID,
-                namespaceID: exec.namespace,
-                stateBuffer: finalStateBuf,
-                logger: exec.logger
-            )
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .workflowCompleted,
-                eventData: nil,
-                logger: exec.logger
-            )
-            return try JSON.encode(output)
-
-        case .failure(let error):
-            // StrandError.cancelled is the internal suspension signal — not a real failure.
-            if case .cancelled? = error as? StrandError { break }
-            // _ContinueAsNewSignal is a lifecycle transition, not a failure.
-            // Re-throw it unwrapped so runTask's typed catch matches it.
-            // Without this guard, `lastCallSite` (set by the preceding
-            // context.runActivity / runChildWorkflow call) would cause
-            // CallSiteAnnotatedError to wrap the signal — defeating the
-            // `catch let signal as _ContinueAsNewSignal` in runTask.
-            if let signal = error as? _ContinueAsNewSignal {
-                cache.remove(claimed.taskID)
-                handlerTask.cancel()
-                executor.cancelPending()
-                executor.drain()  // flush jobs enqueued by cancelPending()
-                throw signal
-            }
-            cache.remove(claimed.taskID)
-            handlerTask.cancel()
-            executor.cancelPending()
-            executor.drain()  // flush jobs enqueued by cancelPending()
-            let errData = try? JSON.encode(["error": String(describing: error)])
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .workflowFailed,
-                eventData: errData,
-                logger: exec.logger
-            )
-            // Stamp the last WorkflowContext call site onto the error.
-            if let site = activation.lastCallSite {
-                throw CallSiteAnnotatedError(
-                    underlying: error,
-                    fileID: site.fileID,
-                    line: site.line
-                )
-            }
-            throw error
-
-        case .none:  // Handler suspended — apply schedule commands.
-            break
-        }
-
-        // ── Suspended: apply schedule commands ────────────────────────────────────
-        let scheduleCommands = commands.filter { cmd in
-            switch cmd {
-            case .scheduleActivity, .startTimer, .awaitEvent, .scheduleChildWorkflow:
-                return true
-            case .writeCheckpoint, .timerFired, .eventReceived:
-                return false
-            }
-        }
-
-        if !scheduleCommands.isEmpty {
-            // Enqueue activities (idempotent — safe to call even if already exists).
-            var enqueued: [(seqNum: Int, taskID: UUID)] = []
-            for cmd in scheduleCommands {
-                switch cmd {
-                case .scheduleActivity(
-                    let name,
-                    let actInput,
-                    let options,
-                    let seqNum,
-                    let idKey
-                ):
-                    let row = try await Queries.enqueueTask(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        queue: options.queue ?? exec.queue,
-                        taskName: name,
-                        paramsBuffer: actInput,
-                        headersBuffer: nil,
-                        retryStrategyBuffer: options.retryStrategy.flatMap { try? JSON.encode($0) },
-                        maxAttempts: options.maxAttempts,
-                        cancellationBuffer: nil,
-                        idempotencyKey: idKey,
-                        priority: options.priority.rawValue,
-                        scheduledAt: options.delayUntil,
-                        timeoutSeconds: options.timeout.map { Int($0.components.seconds) },
-                        deadlineAt: options.maxDuration.map {
-                            Date.now.addingTimeInterval(
-                                Double($0.components.seconds)
-                                    + Double($0.components.attoseconds)
-                                    / 1_000_000_000_000_000_000
-                            )
-                        },
-                        fairnessKey: options.fairnessKey,
-                        fairnessWeight: options.fairnessWeight,
-                        kind: .activity,
-                        parentTaskID: claimed.taskID,
-                        logger: exec.logger
-                    )
-                    enqueued.append((seqNum: seqNum, taskID: row.taskID))
-                    let actData = try? JSON.encode(["activity": name, "seq_num": String(seqNum)])
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .activityScheduled,
-                        eventData: actData,
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .startTimer(let wakeAt, _):
-                    // Transition run to SLEEPING. Both UPDATEs are one SQL statement
-                    // (data-modifying CTE) so PostgreSQL commits them atomically —
-                    // no explicit BEGIN/COMMIT wrapper needed.
-                    try await exec.postgres.query(
-                        """
-                        WITH r AS (
-                            UPDATE strand.runs
-                            SET state = \(TaskState.sleeping), available_at = \(wakeAt),
-                                worker_id = NULL, lease_expires_at = NULL
-                            WHERE id = \(claimed.runID)
-                            RETURNING id
-                        )
-                        UPDATE strand.tasks SET state = \(TaskState.sleeping)
-                        WHERE id = \(claimed.taskID)
-                        """,
-                        logger: exec.logger
-                    )
-                    let timerData = try? JSON.encode([
-                        "duration_ms": Int(wakeAt.timeIntervalSinceNow * 1000)
-                    ])
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .timerStarted,
-                        eventData: timerData,
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .awaitEvent(let eventName, let seqNum, let timeoutAt):
-                    // Atomic check-or-wait: prevents the lost-wakeup race where the
-                    // event fires between drain() returning and this transaction committing.
-                    //
-                    // Without this check:
-                    //   1. drain() returns — handler suspended, .awaitEvent command emitted
-                    //   2. client.emitEvent fires — finds no event_wait yet — run not woken
-                    //   3. worker inserts event_wait — run goes WAITING and stalls forever
-                    //
-                    // Both operations live in the same transaction: insert the event_wait
-                    // and check strand.events. If the event is already there, go PENDING
-                    // immediately so the next activation fast-paths through waitForEvent.
-                    let taskID = claimed.taskID
-                    let runID = claimed.runID
-                    let queueName = exec.queue
-
-                    try await exec.postgres.withTransaction(logger: exec.logger) { conn in
-                        // Always register the event_wait so emitEvent can find this run later.
-                        try await conn.query(
-                            """
-                            INSERT INTO strand.event_waits
-                                (task_id, run_id, queue, seq_num, event_name, timeout_at)
-                            VALUES (\(taskID), \(runID), \(queueName), \(seqNum), \(eventName), \(timeoutAt))
-                            ON CONFLICT (run_id, seq_num) DO UPDATE
-                                SET event_name = EXCLUDED.event_name, timeout_at = EXCLUDED.timeout_at
-                            """,
-                            logger: exec.logger
-                        )
-
-                        // Check if the event was already emitted (race window detection).
-                        let evtStream = try await conn.query(
-                            """
-                            SELECT payload FROM strand.events
-                            WHERE namespace_id = \(exec.namespace)
-                              AND queue = \(queueName)
-                              AND name  = \(eventName)
-                              AND payload IS NOT NULL
-                            """,
-                            logger: exec.logger
-                        )
-
-                        if let evtRow = try await evtStream.first(where: { _ in true }) {
-                            // Event already in strand.events — wake the run immediately.
-                            var col = evtRow.makeIterator()
-                            let existingPayload = try col.next()!.decode(
-                                ByteBuffer?.self,
-                                context: .default
-                            )
-
-                            // Transition to PENDING with the event payload so the next
-                            // activation's waitForEvent fast-path (wakeEvent == name) fires.
-                            try await conn.query(
-                                """
-                                WITH r AS (
-                                    UPDATE strand.runs
-                                    SET state        = \(TaskState.pending),
-                                        available_at  = NOW(),
-                                        wake_event    = \(eventName),
-                                        event_payload = \(existingPayload),
-                                        worker_id     = NULL,
-                                        lease_expires_at = NULL
-                                    WHERE id = \(runID)
-                                )
-                                UPDATE strand.tasks SET state = \(TaskState.pending)
-                                WHERE id = \(taskID)
-                                """,
-                                logger: exec.logger
-                            )
-                        } else {
-                            // Event not yet emitted — set run to WAITING (or SLEEPING for timed waits).
-                            let availableAt =
-                                timeoutAt ?? Date(timeIntervalSince1970: 32_503_680_000)
-                            let runState: TaskState = timeoutAt != nil ? .sleeping : .waiting
-                            try await conn.query(
-                                """
-                                WITH r AS (
-                                    UPDATE strand.runs
-                                    SET state        = \(runState),
-                                        available_at  = \(availableAt),
-                                        wake_event    = \(eventName),
-                                        event_payload = NULL,
-                                        worker_id     = NULL,
-                                        lease_expires_at = NULL
-                                    WHERE id = \(runID)
-                                )
-                                UPDATE strand.tasks SET state = \(runState)
-                                WHERE id = \(taskID)
-                                """,
-                                logger: exec.logger
-                            )
-                        }
-                    }
-                    let eventWaitData = try? JSON.encode(["event_name": eventName])
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .eventWaitStarted,
-                        eventData: eventWaitData,
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .scheduleChildWorkflow(
-                    let name,
-                    let childQueue,
-                    let wfInput,
-                    let seqNum,
-                    let idKey
-                ):
-                    let targetQueue = childQueue ?? exec.queue  // use child's queue or inherit
-                    let row = try await Queries.enqueueTask(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        queue: targetQueue,
-                        taskName: name,
-                        paramsBuffer: wfInput,
-                        headersBuffer: nil,
-                        retryStrategyBuffer: nil,
-                        maxAttempts: nil,
-                        cancellationBuffer: nil,
-                        idempotencyKey: idKey,
-                        priority: TaskPriority.normal.rawValue,
-                        scheduledAt: nil,
-                        fairnessKey: nil,
-                        fairnessWeight: 1.0,
-                        kind: .workflow,
-                        parentTaskID: claimed.taskID,
-                        logger: exec.logger
-                    )
-                    enqueued.append((seqNum: seqNum, taskID: row.taskID))
-                    let childData = try? JSON.encode([
-                        "workflow": name, "seq_num": String(seqNum),
-                    ])
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .childWorkflowStarted,
-                        eventData: childData,
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .writeCheckpoint, .timerFired, .eventReceived:
-                    break  // already handled above
-                }
-            }
-
-            // Register event_waits for all enqueued activities / child workflows,
-            // then transition the run to WAITING (event-driven, no timer).
-            //
-            // Atomic check-or-wait: prevents the lost-wakeup race where a child
-            // task completes between enqueueTask() committing and this transaction
-            // committing.
-            //
-            // Without this check:
-            //   1. enqueueTask() commits — child task visible to activity workers
-            //   2. Activity worker claims, runs, and completes it immediately
-            //   3. emitTaskCompletionSignal finds no event_wait — parent not woken
-            //   4. This transaction commits — run goes WAITING and stalls forever
-            //
-            // Both the event_wait inserts and a task_completions count live in the
-            // same transaction. If all children are already done, go PENDING immediately;
-            // the next poll re-activates the workflow. Remaining event_waits fire normally.
-            if !enqueued.isEmpty {
-                try await exec.postgres.withTransaction(logger: exec.logger) { conn in
-                    // TODO: optimize this path
-                    for item in enqueued {
-                        try await conn.query(
-                            """
-                            INSERT INTO strand.event_waits
-                                (task_id, run_id, queue, seq_num, child_task_id, timeout_at)
-                            VALUES (\(claimed.taskID), \(claimed.runID), \(exec.queue),
-                                    \(item.seqNum), \(item.taskID), NULL)
-                            ON CONFLICT (run_id, seq_num) DO UPDATE
-                                SET child_task_id = EXCLUDED.child_task_id,
-                                    timeout_at    = NULL
-                            """,
-                            logger: exec.logger
-                        )
-                    }
-
-                    // Atomically check whether any child tasks already completed
-                    // (i.e. are present in task_completions) before we park the run.
-                    let childTaskIDs = enqueued.map { $0.taskID }
-                    let completedStream = try await conn.query(
-                        """
-                        SELECT COUNT(*) FROM strand.task_completions
-                        WHERE task_id = ANY(\(childTaskIDs))
-                        """,
-                        logger: exec.logger
-                    )
-                    var completedCount = 0
-                    for try await row in completedStream {
-                        var col = row.makeIterator()
-                        completedCount = try col.next()!.decode(Int.self, context: .default)
-                    }
-
-                    if completedCount == enqueued.count {
-                        // All children already finished — go PENDING so the next poll
-                        // re-activates immediately rather than waiting for events that
-                        // already fired and were missed.
-                        try await conn.query(
-                            """
-                            WITH r AS (
-                                UPDATE strand.runs
-                                SET state = \(TaskState.pending), available_at = NOW(),
-                                    worker_id = NULL, lease_expires_at = NULL
-                                WHERE id = \(claimed.runID)
-                                RETURNING id
-                            )
-                            UPDATE strand.tasks SET state = \(TaskState.pending)
-                            WHERE id = \(claimed.taskID)
-                            """,
-                            logger: exec.logger
-                        )
-                    } else {
-                        try await conn.query(
-                            """
-                            WITH r AS (
-                                UPDATE strand.runs
-                                SET state = \(TaskState.waiting), worker_id = NULL, lease_expires_at = NULL
-                                WHERE id = \(claimed.runID)
-                                RETURNING id
-                            )
-                            UPDATE strand.tasks SET state = \(TaskState.waiting)
-                            WHERE id = \(claimed.taskID)
-                            """,
-                            logger: exec.logger
-                        )
-                    }
-                }
-            }
-        }
-
-        // ── Unsatisfied conditions ──────────────────────────────────────────────────
-        // Any conditions that were not satisfied post-drain need a DB state transition
-        // so the worker releases this run until a signal or timer wakes it.
-        //
-        // • condition(_:)          → WAITING (no timer; woken only by a signal)
-        // • condition(_:timeout:)  → SLEEPING with available_at = deadline
-        //   (woken by timer expiry OR by a signal that transitions SLEEPING → PENDING)
-        if executor.hasUnsatisfiedConditions && scheduleCommands.isEmpty {
-            let condTaskID = claimed.taskID
-            let condRunID = claimed.runID
-            if let wakeAt = executor.conditionMinWakeAt {
-                // At least one condition has a deadline — sleep until the earliest one.
-                try await exec.postgres.query(
-                    """
-                    WITH r AS (
-                        UPDATE strand.runs
-                        SET state = \(TaskState.sleeping), available_at = \(wakeAt),
-                            worker_id = NULL, lease_expires_at = NULL
-                        WHERE id = \(condRunID)
-                    )
-                    UPDATE strand.tasks SET state = \(TaskState.sleeping) WHERE id = \(condTaskID)
-                    """,
-                    logger: exec.logger
-                )
-            } else {
-                // All conditions are indefinite — wait for a signal.
-                try await exec.postgres.query(
-                    """
-                    WITH r AS (
-                        UPDATE strand.runs
-                        SET state = \(TaskState.waiting), worker_id = NULL, lease_expires_at = NULL
-                        WHERE id = \(condRunID)
-                    )
-                    UPDATE strand.tasks SET state = \(TaskState.waiting) WHERE id = \(condTaskID)
-                    """,
-                    logger: exec.logger
-                )
-            }
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .conditionWaiting,
-                eventData: nil,
-                logger: exec.logger
-            )
-            historySeq += 1
-        }
-        // No else branch: if scheduleCommands is empty and no conditions remain,
-        // suspension came from sleep/waitForEvent which already wrote their own DB updates.
-
-        // ── 8. Cache the live Task for the next activation ────────────────────────────
-        // The handler is parked on a continuation — do NOT call cancelPending().
-        // On the next activation, resumeActivation() will deliver real results
-        // directly to those continuations and call drain() to continue execution.
-        cache.set(
-            claimed.taskID,
-            _WorkflowTaskCache<W>.CachedState(
-                executor: executor,
-                task: handlerTask,
-                handlerResult: handlerResult,
-                stateBox: stateBox,
-                activation: activation
-            )
-        )
-
-        return nil
-    }
-
-    // MARK: - Cached re-activation
-
-    /// Re-activates a workflow whose handler Task is still alive and parked on
-    /// continuations from the previous activation.
-    ///
-    /// Instead of replaying the handler from scratch this method:
-    ///   1. Applies any pending signals to the live `stateBox`.
-    ///   2. Writes checkpoints for newly completed child activities.
-    ///   3. Resumes the specific continuations that now have results.
-    ///   4. Calls `drain()` to let the handler continue from where it paused.
-    ///   5. Applies new schedule commands to Postgres (identical to fresh path).
-    private func resumeActivation(
-        claimed: ClaimedTask,
-        exec: _WorkerExec,
-        cached: _WorkflowTaskCache<W>.CachedState,
-        cache: _WorkflowTaskCache<W>
-    ) async throws -> ByteBuffer? {
-
-        let executor = cached.executor
-        let handlerResult = cached.handlerResult
-        let stateBox = cached.stateBox
-        let activation = cached.activation
-        let handlerTask = cached.task
-        let claimTimeoutSecs = Int(exec.options.claimTimeout.components.seconds)
-
-        // ── History sequence ──────────────────────────────────────────────────
-        var historySeq = try await WorkflowStateQueries.nextHistorySeq(
-            on: exec.postgres,
-            taskID: claimed.taskID,
-            logger: exec.logger
-        )
-
-        // ── External checkpoint refresh ──────────────────────────────────────────────
-        // `client.markVersion` writes checkpoints directly to the DB without sending
-        // a signal. It is only meaningful during a workflow sleep (timer wait), so we
-        // only scan for external writes when the run was NOT woken by a child-workflow
-        // or event completion (those set wake_event; timers and signals leave it nil).
-        // On the hot path (child completion, wake_event != nil) we skip the scan
-        // entirely — eliminating a full checkpoint-table roundtrip per child.
-        if claimed.wakeEvent == nil {
-            let freshCheckpoints = try await Queries.getCheckpointStates(
-                on: exec.postgres,
-                taskID: claimed.taskID,
-                runID: claimed.runID,
-                logger: exec.logger
-            )
-            for row in freshCheckpoints where activation.cachedCheckpoint(for: row.seqNum) == nil {
-                activation.cacheCheckpoint(seqNum: row.seqNum, buffer: row.stateBuffer)
-            }
-        }
-
-        // ── Signals ──────────────────────────────────────────────────────────
-        // Apply directly to the live stateBox — no need to reload from DB.
-        let signals = try await WorkflowStateQueries.loadPendingSignals(
-            on: exec.postgres,
-            taskID: claimed.taskID,
-            namespaceID: exec.namespace,
-            logger: exec.logger
-        )
-        for signal in signals {
-            try stateBox.value.handleSignal(name: signal.name, payload: signal.payload)
-        }
-        if !signals.isEmpty {
-            for signal in signals {
-                let sigData = try? JSON.encode(["name": signal.name])
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .signalReceived,
-                    eventData: sigData,
-                    logger: exec.logger
-                )
-                historySeq += 1
-            }
-            let postSignalBuf = try JSON.encode(stateBox.value)
-            try await WorkflowStateQueries.saveState(
-                on: exec.postgres,
-                taskID: claimed.taskID,
-                namespaceID: exec.namespace,
-                stateBuffer: postSignalBuf,
-                logger: exec.logger
-            )
-            try await WorkflowStateQueries.deleteSignals(
-                on: exec.postgres,
-                ids: signals.map { $0.id },
-                logger: exec.logger
-            )
-
-        }
-
-        // ── Completed children ──────────────────────────────────────────────────
-        let completedChildren = try await WorkflowStateQueries.loadCompletedChildActivities(
-            on: exec.postgres,
-            parentTaskID: claimed.taskID,
-            logger: exec.logger
-        )
-
-        // Child-activity results are already durable in strand.task_completions.
-        // The crash-recovery (fresh) path reads them from there via fast path 2
-        // (resolveCompleted → preloadedResults), so writing them again to
-        // strand.checkpoints is redundant.  We still extend the claim so that
-        // activations processing many simultaneous completions don’t expire.
-        if !completedChildren.isEmpty {
-            try await Queries.extendClaim(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                runID: claimed.runID,
-                extendBySeconds: claimTimeoutSecs,
-                logger: exec.logger
-            )
-        }
-
-        for (seqNum, _, _, state, kind, name) in completedChildren
-        where kind == .workflow && state == .completed {
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .childWorkflowCompleted,
-                eventData: try? JSON.encode(["workflow": name, "seq_num": String(seqNum)]),
-                logger: exec.logger
-            )
-            historySeq += 1
-        }
-
-        // ── Event/timer checkpoint (before resuming, for durability) ──────────
-        if let eventName = claimed.wakeEvent,
-            let seqNum = executor.seqNum(forEventName: eventName),
-            activation.cachedCheckpoint(for: seqNum) == nil
-        {
-            if let payload = claimed.eventPayloadBuffer {
-                try await Queries.batchSetCheckpoints(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    runID: claimed.runID,
-                    checkpoints: [(seqNum: seqNum, name: "waitForEvent:\(eventName)", state: payload)],
-                    extendClaimBySeconds: claimTimeoutSecs,
-                    logger: exec.logger
-                )
-                activation.cacheCheckpoint(seqNum: seqNum, buffer: payload)
-            } else {
-                let sentinel = try JSON.encode(TimeoutSentinel())
-                try await Queries.batchSetCheckpoints(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    runID: claimed.runID,
-                    checkpoints: [(seqNum: seqNum, name: "waitForEvent:\(eventName):timeout", state: sentinel)],
-                    extendClaimBySeconds: claimTimeoutSecs,
-                    logger: exec.logger
-                )
-                activation.cacheCheckpoint(seqNum: seqNum, buffer: sentinel)
-            }
-        }
-
-        // ── Resume parked continuations ───────────────────────────────────────
-        // Clear commands from the previous activation before re-draining.
-        executor.clearPendingCommands()
-
-        for (seqNum, result, failureReason, state, kind, name) in completedChildren {
-            switch state {
-            case .completed:
-                if let result { executor.resumeActivity(seqNum: seqNum, result: result) }
-            case .failed, .cancelled:
-                let retryState: ActivityRetryState =
-                    state == .cancelled ? .cancelled : .maximumAttemptsReached
-                let err: Error
-                if kind == .workflow {
-                    err = StrandError.childWorkflowFailed(name: name, state: state.rawValue)
-                } else {
-                    // Use _ActivityFailureSignal so runActivity<A> can decode A.Failure.
-                    err = _ActivityFailureSignal(
-                        failureReason: failureReason,
-                        retryState: retryState,
-                        activityName: name,
-                        state: state
-                    )
-                }
-                executor.resumeActivityFailure(seqNum: seqNum, error: err)
-            default:
-                break
-            }
-        }
-
-        if let eventName = claimed.wakeEvent,
-            let seqNum = executor.seqNum(forEventName: eventName)
-        {
-            if let payload = claimed.eventPayloadBuffer {
-                executor.resumeEvent(seqNum: seqNum, payload: payload)
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .eventReceived,
-                    eventData: try? JSON.encode(["event_name": eventName]),
-                    logger: exec.logger
-                )
-                historySeq += 1
-            } else {
-                executor.resumeEventWithTimeout(seqNum: seqNum, eventName: eventName)
-            }
-        }
-
-        // If woken without an event name a sleep timer (or condition deadline) fired.
-        if claimed.wakeEvent == nil { executor.resumeAllTimers() }
-
-        // ── Drain + conditions ────────────────────────────────────────────────
-        executor.drain()
-        while executor.resumeExpiredConditions() { executor.drain() }
-        while executor.evaluateAndResumeFirstSatisfiedCondition() { executor.drain() }
-
-        // ── Local activity execution loop ─────────────────────────────────────
-        localActivityLoop: while !executor.localActivityEntries.isEmpty {
-            let entries = executor.localActivityEntries
-            for (id, entry) in entries {
-                guard let runner = exec.localActivityLookup[entry.name] else {
-                    executor.failLocalActivity(id: id, error: StrandError.unknownTask(name: entry.name))
-                    continue
-                }
-                do {
-                    let result = try await runner(entry.input, exec, claimed.taskID)
-                    try await Queries.setCheckpointState(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seqNum: entry.seqNum,
-                        name: entry.name,
-                        stateBuffer: result,
-                        runID: claimed.runID,
-                        extendClaimBySeconds: claimTimeoutSecs,
-                        logger: exec.logger
-                    )
-                    activation.cacheCheckpoint(seqNum: entry.seqNum, buffer: result)
-                    executor.resolveLocalActivity(id: id, result: result)
-                } catch {
-                    executor.failLocalActivity(id: id, error: error)
-                }
-            }
-            executor.drain()
-            while executor.resumeExpiredConditions() { executor.drain() }
-            while executor.evaluateAndResumeFirstSatisfiedCondition() { executor.drain() }
-            if handlerResult.value != nil { break localActivityLoop }
-        }
-
-        // ── Apply new commands to Postgres ────────────────────────────────────
-        let commands = executor.pendingCommands
-
-        let checkpointWrites = commands.compactMap {
-            cmd -> (seqNum: Int, name: String?, state: ByteBuffer)? in
-            if case .writeCheckpoint(let seqNum, let name, let value) = cmd {
-                return (seqNum: seqNum, name: name, state: value)
-            }
-            return nil
-        }
-        if !checkpointWrites.isEmpty {
-            try await Queries.batchSetCheckpoints(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                runID: claimed.runID,
-                checkpoints: checkpointWrites,
-                extendClaimBySeconds: claimTimeoutSecs,
-                logger: exec.logger
-            )
-            for (seqNum, _, value) in checkpointWrites {
-                activation.cacheCheckpoint(seqNum: seqNum, buffer: value)
-            }
-        }
-
-        for cmd in commands {
-            switch cmd {
-            case .timerFired:
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .timerFired,
-                    eventData: nil,
-                    logger: exec.logger
-                )
-                historySeq += 1
-            case .eventReceived(let evtName):
-                try await WorkflowStateQueries.appendHistory(
-                    on: exec.postgres,
-                    namespaceID: exec.namespace,
-                    taskID: claimed.taskID,
-                    seq: historySeq,
-                    eventType: .eventReceived,
-                    eventData: try? JSON.encode(["event_name": evtName]),
-                    logger: exec.logger
-                )
-                historySeq += 1
-            default:
-                break
-            }
-        }
-
-        // ── Result handling ───────────────────────────────────────────────────
-        switch handlerResult.value {
-
-        case .success(let output):
-            cache.remove(claimed.taskID)
-            handlerTask.cancel()
-            executor.cancelPending()
-            executor.drain()  // flush residual jobs from cancelPending()
-            let finalStateBuf = try JSON.encode(stateBox.value)
-            try await WorkflowStateQueries.saveState(
-                on: exec.postgres,
-                taskID: claimed.taskID,
-                namespaceID: exec.namespace,
-                stateBuffer: finalStateBuf,
-                logger: exec.logger
-            )
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .workflowCompleted,
-                eventData: nil,
-                logger: exec.logger
-            )
-            return try JSON.encode(output)
-
-        case .failure(let error):
-            if case .cancelled? = error as? StrandError { break }
-            // Same unwrap-before-wrap guard as in _activate: re-throw _ContinueAsNewSignal
-            // directly so runTask's typed catch matches it (see _activate for rationale).
-            if let signal = error as? _ContinueAsNewSignal {
-                cache.remove(claimed.taskID)
-                handlerTask.cancel()
-                executor.cancelPending()
-                executor.drain()  // flush residual jobs from cancelPending()
-                throw signal
-            }
-            cache.remove(claimed.taskID)
-            handlerTask.cancel()
-            executor.cancelPending()
-            executor.drain()  // flush residual jobs from cancelPending()
-            let errData = try? JSON.encode(["error": String(describing: error)])
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .workflowFailed,
-                eventData: errData,
-                logger: exec.logger
-            )
-            if let site = activation.lastCallSite {
-                throw CallSiteAnnotatedError(underlying: error, fileID: site.fileID, line: site.line)
-            }
-            throw error
-
-        case .none:
-            break  // Handler still suspended — fall through to schedule commands
-        }
-
-        // ── Suspended: schedule new activities / timers / events ─────────────
-        let scheduleCommands = commands.filter { cmd in
-            switch cmd {
-            case .scheduleActivity, .startTimer, .awaitEvent, .scheduleChildWorkflow: return true
-            case .writeCheckpoint, .timerFired, .eventReceived: return false
-            }
-        }
-
-        if !scheduleCommands.isEmpty {
-            var enqueued: [(seqNum: Int, taskID: UUID)] = []
-            for cmd in scheduleCommands {
-                switch cmd {
-                case .scheduleActivity(let name, let actInput, let options, let seqNum, let idKey):
-                    let row = try await Queries.enqueueTask(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        queue: options.queue ?? exec.queue,
-                        taskName: name,
-                        paramsBuffer: actInput,
-                        headersBuffer: nil,
-                        retryStrategyBuffer: options.retryStrategy.flatMap { try? JSON.encode($0) },
-                        maxAttempts: options.maxAttempts,
-                        cancellationBuffer: nil,
-                        idempotencyKey: idKey,
-                        priority: options.priority.rawValue,
-                        scheduledAt: options.delayUntil,
-                        timeoutSeconds: options.timeout.map { Int($0.components.seconds) },
-                        deadlineAt: options.maxDuration.map {
-                            Date.now.addingTimeInterval(
-                                Double($0.components.seconds)
-                                    + Double($0.components.attoseconds) / 1_000_000_000_000_000_000
-                            )
-                        },
-                        fairnessKey: options.fairnessKey,
-                        fairnessWeight: options.fairnessWeight,
-                        kind: .activity,
-                        parentTaskID: claimed.taskID,
-                        logger: exec.logger
-                    )
-                    enqueued.append((seqNum: seqNum, taskID: row.taskID))
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .activityScheduled,
-                        eventData: try? JSON.encode(["activity": name, "seq_num": String(seqNum)]),
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .startTimer(let wakeAt, _):
-                    try await exec.postgres.query(
-                        """
-                        WITH r AS (
-                            UPDATE strand.runs
-                            SET state = \(TaskState.sleeping), available_at = \(wakeAt),
-                                worker_id = NULL, lease_expires_at = NULL
-                            WHERE id = \(claimed.runID)
-                            RETURNING id
-                        )
-                        UPDATE strand.tasks SET state = \(TaskState.sleeping)
-                        WHERE id = \(claimed.taskID)
-                        """,
-                        logger: exec.logger
-                    )
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .timerStarted,
-                        eventData: try? JSON.encode(["duration_ms": Int(wakeAt.timeIntervalSinceNow * 1000)]),
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .awaitEvent(let evtName, let seqNum, let timeoutAt):
-                    let taskID = claimed.taskID
-                    let runID = claimed.runID
-                    let queueName = exec.queue
-                    try await exec.postgres.withTransaction(logger: exec.logger) { conn in
-                        try await conn.query(
-                            """
-                            INSERT INTO strand.event_waits
-                                (task_id, run_id, queue, seq_num, event_name, timeout_at)
-                            VALUES (\(taskID), \(runID), \(queueName), \(seqNum), \(evtName), \(timeoutAt))
-                            ON CONFLICT (run_id, seq_num) DO UPDATE
-                                SET event_name = EXCLUDED.event_name, timeout_at = EXCLUDED.timeout_at
-                            """,
-                            logger: exec.logger
-                        )
-                        let evtStream = try await conn.query(
-                            """
-                            SELECT payload FROM strand.events
-                            WHERE namespace_id = \(exec.namespace)
-                              AND queue        = \(queueName)
-                              AND name         = \(evtName)
-                              AND payload IS NOT NULL
-                            """,
-                            logger: exec.logger
-                        )
-                        if let evtRow = try await evtStream.first(where: { _ in true }) {
-                            var col = evtRow.makeIterator()
-                            let existing = try col.next()!.decode(ByteBuffer?.self, context: .default)
-                            try await conn.query(
-                                """
-                                WITH r AS (
-                                    UPDATE strand.runs
-                                    SET state = \(TaskState.pending), available_at = NOW(),
-                                        wake_event = \(evtName), event_payload = \(existing),
-                                        worker_id = NULL, lease_expires_at = NULL
-                                    WHERE id = \(runID)
-                                )
-                                UPDATE strand.tasks SET state = \(TaskState.pending)
-                                WHERE id = \(taskID)
-                                """,
-                                logger: exec.logger
-                            )
-                        } else {
-                            let availableAt = timeoutAt ?? Date(timeIntervalSince1970: 32_503_680_000)
-                            let runState: TaskState = timeoutAt != nil ? .sleeping : .waiting
-                            try await conn.query(
-                                """
-                                WITH r AS (
-                                    UPDATE strand.runs
-                                    SET state = \(runState), available_at = \(availableAt),
-                                        wake_event = \(evtName), event_payload = NULL,
-                                        worker_id = NULL, lease_expires_at = NULL
-                                    WHERE id = \(runID)
-                                )
-                                UPDATE strand.tasks SET state = \(runState)
-                                WHERE id = \(taskID)
-                                """,
-                                logger: exec.logger
-                            )
-                        }
-                    }
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .eventWaitStarted,
-                        eventData: try? JSON.encode(["event_name": evtName]),
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .scheduleChildWorkflow(let name, let childQueue, let wfInput, let seqNum, let idKey):
-                    let row = try await Queries.enqueueTask(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        queue: childQueue ?? exec.queue,
-                        taskName: name,
-                        paramsBuffer: wfInput,
-                        headersBuffer: nil,
-                        retryStrategyBuffer: nil,
-                        maxAttempts: nil,
-                        cancellationBuffer: nil,
-                        idempotencyKey: idKey,
-                        priority: TaskPriority.normal.rawValue,
-                        scheduledAt: nil,
-                        fairnessKey: nil,
-                        fairnessWeight: 1.0,
-                        kind: .workflow,
-                        parentTaskID: claimed.taskID,
-                        logger: exec.logger
-                    )
-                    enqueued.append((seqNum: seqNum, taskID: row.taskID))
-                    try await WorkflowStateQueries.appendHistory(
-                        on: exec.postgres,
-                        namespaceID: exec.namespace,
-                        taskID: claimed.taskID,
-                        seq: historySeq,
-                        eventType: .childWorkflowStarted,
-                        eventData: try? JSON.encode(["workflow": name, "seq_num": String(seqNum)]),
-                        logger: exec.logger
-                    )
-                    historySeq += 1
-
-                case .writeCheckpoint, .timerFired, .eventReceived:
-                    break
-                }
-            }
-
-            if !enqueued.isEmpty {
-                try await exec.postgres.withTransaction(logger: exec.logger) { conn in
-                    for item in enqueued {
-                        try await conn.query(
-                            """
-                            INSERT INTO strand.event_waits
-                                (task_id, run_id, queue, seq_num, child_task_id, timeout_at)
-                            VALUES (\(claimed.taskID), \(claimed.runID), \(exec.queue),
-                                    \(item.seqNum), \(item.taskID), NULL)
-                            ON CONFLICT (run_id, seq_num) DO UPDATE
-                                SET child_task_id = EXCLUDED.child_task_id,
-                                    timeout_at    = NULL
-                            """,
-                            logger: exec.logger
-                        )
-                    }
-                    let childTaskIDs = enqueued.map { $0.taskID }
-                    let completedStream = try await conn.query(
-                        "SELECT COUNT(*) FROM strand.task_completions WHERE task_id = ANY(\(childTaskIDs))",
-                        logger: exec.logger
-                    )
-                    var completedCount = 0
-                    for try await row in completedStream {
-                        var col = row.makeIterator()
-                        completedCount = try col.next()!.decode(Int.self, context: .default)
-                    }
-                    if completedCount == enqueued.count {
-                        try await conn.query(
-                            """
-                            WITH r AS (
-                                UPDATE strand.runs
-                                SET state = \(TaskState.pending), available_at = NOW(),
-                                    worker_id = NULL, lease_expires_at = NULL
-                                WHERE id = \(claimed.runID) RETURNING id
-                            )
-                            UPDATE strand.tasks SET state = \(TaskState.pending) WHERE id = \(claimed.taskID)
-                            """,
-                            logger: exec.logger
-                        )
-                    } else {
-                        try await conn.query(
-                            """
-                            WITH r AS (
-                                UPDATE strand.runs
-                                SET state = \(TaskState.waiting), worker_id = NULL, lease_expires_at = NULL
-                                WHERE id = \(claimed.runID) RETURNING id
-                            )
-                            UPDATE strand.tasks SET state = \(TaskState.waiting) WHERE id = \(claimed.taskID)
-                            """,
-                            logger: exec.logger
-                        )
-                    }
-                }
-            }
-        }
-
-        if executor.hasUnsatisfiedConditions && scheduleCommands.isEmpty {
-            if let wakeAt = executor.conditionMinWakeAt {
-                try await exec.postgres.query(
-                    """
-                    WITH r AS (
-                        UPDATE strand.runs
-                        SET state = \(TaskState.sleeping), available_at = \(wakeAt),
-                            worker_id = NULL, lease_expires_at = NULL
-                        WHERE id = \(claimed.runID)
-                    )
-                    UPDATE strand.tasks SET state = \(TaskState.sleeping) WHERE id = \(claimed.taskID)
-                    """,
-                    logger: exec.logger
-                )
-            } else {
-                try await exec.postgres.query(
-                    """
-                    WITH r AS (
-                        UPDATE strand.runs
-                        SET state = \(TaskState.waiting), worker_id = NULL, lease_expires_at = NULL
-                        WHERE id = \(claimed.runID)
-                    )
-                    UPDATE strand.tasks SET state = \(TaskState.waiting) WHERE id = \(claimed.taskID)
-                    """,
-                    logger: exec.logger
-                )
-            }
-            try await WorkflowStateQueries.appendHistory(
-                on: exec.postgres,
-                namespaceID: exec.namespace,
-                taskID: claimed.taskID,
-                seq: historySeq,
-                eventType: .conditionWaiting,
-                eventData: nil,
-                logger: exec.logger
-            )
-        }
-
-        // ── Partial-completion re-wait ──────────────────────────────────────────────────
-        // The handler resumed some (but not all) children in a `withThrowingTaskGroup`
-        // batch and is parked waiting for the rest. No new schedule commands were
-        // emitted. Transition to WAITING so the next child completion can wake this
-        // run through the `event_waits` registered during the original dispatch.
-        if executor.hasPendingContinuations
-            && scheduleCommands.isEmpty
-            && !executor.hasUnsatisfiedConditions
-        {
-            // Single CTE: transition the run to WAITING (so the next child
-            // completion can wake it), but check atomically whether any
-            // event_waits for this run ALREADY have a matching task_completion
-            // (meaning the child completed while we were RUNNING and its
-            // event_wait survived because the wake was a no-op). If so, go
-            // straight to PENDING — no need to wait for another signal.
-            try await exec.postgres.query(
-                """
-                WITH
-                missed AS (
-                    -- event_waits whose children have already completed but whose
-                    -- wake signal was silently dropped (parent was RUNNING at the time).
-                    -- child_task_id is a typed UUID FK — direct join, no string parsing.
-                    SELECT COUNT(*) AS cnt
-                    FROM strand.event_waits ew
-                    JOIN strand.task_completions tc ON tc.task_id = ew.child_task_id
-                    WHERE ew.run_id        = \(claimed.runID)
-                      AND ew.child_task_id IS NOT NULL
-                ),
-                run_upd AS (
-                    UPDATE strand.runs
-                    SET state            = CASE WHEN (SELECT cnt FROM missed) > 0
-                                               THEN \(TaskState.pending)
-                                               ELSE \(TaskState.waiting) END,
-                        available_at     = CASE WHEN (SELECT cnt FROM missed) > 0
-                                               THEN NOW()
-                                               ELSE available_at END,
-                        worker_id        = NULL,
-                        lease_expires_at = NULL
-                    WHERE id = \(claimed.runID)
-                    RETURNING state
-                )
-                UPDATE strand.tasks
-                SET state = (SELECT state FROM run_upd)
-                WHERE id = \(claimed.taskID)
-                  AND namespace_id = \(exec.namespace)
-                """,
-                logger: exec.logger
-            )
-        }
-
-        // Still suspended — keep the cached Task alive.
-        return nil
-    }
-}
-
 // MARK: - AnyRegistration
 
 /// Type-erased handler entry stored in ``Registry``.
@@ -2091,8 +537,19 @@ public struct StrandWorker: Service {
                             .recordNanoseconds(elapsed.nanoseconds)
                     }
                     // OTel span: one span per task execution attempt.
-                    // If no tracing backend is bootstrapped this is a zero-cost no-op.
-                    return try await withSpan(claimed.taskName, ofKind: .internal) { span in
+                    // SpanKind.consumer: the worker is processing an async message
+                    // from the task queue. The producer span (enqueue call-site) ended
+                    // before this span starts, so we use addLink rather than
+                    // parent-child — the async gap breaks the parent-child model.
+                    // Zero-cost no-op when no tracing backend is bootstrapped.
+                    var linkCtx = ServiceContext.topLevel
+                    InstrumentationSystem.tracer.extract(
+                        claimed.headers,
+                        into: &linkCtx,
+                        using: DictionaryExtractor()
+                    )
+                    return try await withSpan(claimed.taskName, ofKind: .consumer) { span in
+                        span.addLink(SpanLink(context: linkCtx, attributes: [:]))
                         span.attributes[StrandLogKeys.taskName] = SpanAttribute.string(
                             claimed.taskName
                         )
